TS: IntegratedFilter
